error solved

+++ src/app/modules/user/user.service.ts
@@ -1,10 +1,12 @@
-import { TUser } from './user.interface';
+import { TOrders, TUser } from './user.interface';
 import { User } from './user.model';
 
 const createUser = async (userData: TUser) => {
-  const result = await User.create(userData); //////built in static method
-
-  return result;
+  const result = await User.create(userData);
+  const result1 = await User.findOne({ userId: result.userId }).select(
+    '-orders -password _id -__v',
+  );
+  return result1;
 };
 
 const getAllUsers = async () => {
@@ -25,8 +27,13 @@ const getAllUsers = async () => {
 
 const getSingleUser = async (id: number) => {
   const user = await User.isUserExists(id);
-
-  return user;
+  if (!user) {
+    throw new Error('User not found');
+  }
+  const result = await User.findOne({ userId: id }).select(
+    '-orders -password _id -__v',
+  );
+  return result;
 };
 
 const updateUser = async (id: number, userData: TUser) => {
@@ -44,11 +51,15 @@ const updateUser = async (id: number, userData: TUser) => {
 };
 
 const deleteUser = async (id: number) => {
+  const user = await User.isUserExists(id);
+  if (!user) {
+    throw new Error('User not found');
+  }
   const result = await User.deleteOne({ userId: id });
   return result;
 };
 
-const createOrder = async (id: number, product: TUser) => {
+const createOrder = async (id: number, product: TOrders) => {
   const user = await User.isUserExists(id);
   if (!user) {
     throw new Error('User not found');
@@ -63,12 +74,20 @@ const createOrder = async (id: number, product: TUser) => {
   return result;
 };
 const getAllOrders = async (id: number) => {
+  const user = await User.isUserExists(id);
+  if (!user) {
+    throw new Error('User not found');
+  }
   const result = await User.findOne({ userId: id }, { orders: 1, _id: 0 });
 
   return result;
 };
 
 const getTotalPrice = async (id: number): Promise<TUser | null> => {
+  const user = await User.isUserExists(id);
+  if (!user) {
+    throw new Error('User not found');
+  }
   const result = await User.findOne(
     { userId: id },
     { totalPrice: { $sum: '$orders.price' }, _id: 0 },
